Add analytics_cookies_accepted? helper method
Add #analytics_cookies_accepted? helper method, to workaround a haml-lint issue of using instance variables in partials - InstanceVariables.

This is because Turbo does not refresh the whole page and thus @analytics_cookies_accepted passed in as a local partial variable will not refresh either.
This leads to google analytics cookies not being set when cookies are accepted but the page isn't refreshed.