Clean up JudicialResultsDecorator
Seperate large chunks of functionality in a public method to their own private  methods.

This makes the code easier to read and easier to notice where code can be reused. Also less chance of a change breaking other
functionalities.

Delete the JudicialResultsPromptDecorator from the file as it's now in a seperate file for standardisation purposes.

Update the unit test to call the new method name