Add JS package babel/plugin to fix yarn error
Add  @babel/plugin-proposal-private-property-in-object to fix the below error after running yarn:

Error: [BABEL]: --- PLACEHOLDER PACKAGE ---

This @babel/plugin-proposal-private-property-in-object version is not meant to
be imported. Something is importing
@babel/plugin-proposal-private-property-in-object without declaring it in its
dependencies (or devDependencies) in the package.json file.
Add "@babel/plugin-proposal-private-property-in-object" to your devDependencies
to work around this error. This will make this message go away.