Fix handling of error_message during unlinking exception
Fix the parsing of the exception's error messages.

When a ServerError/ClientError occurs on the defendant unlinking feature, we weren't providing the error message
provided by CDA but instead sending our generic one. I presume the code was trying to show the error message that
we get from CDA but failed to parse it.