Upgrade json_api_client
The new version of json_api_client removes faraday_middleware to enable
compatibility with both version 1 and 2 of Faraday. However, faraday_middleware
is still required for the FaradaMiddleware::OAuth2. This causes conflicts as
json_api_client attempts to load middleware for JSON that exists both in the
faraday and faraday_middleware gems. Version 1.2.1 of the faraday_middleware
fixes this and ensures that the built in version of middleware is favoured.